@@ -4,31 +4,30 @@ build-backend = "hatchling.build"
 
 [project]
 name = "sonos-lastfm"
-version = "0.1.5"
+version = "0.1.7"
 description = "Automatically scrobble music playing on your Sonos speakers to Last.fm"
 readme = "README.md"
 authors = [
     { name = "Denis Moskalets", email = "denya.msk@gmail.com" }
 ]
 license = { text = "MIT" }
-requires-python = ">=3.11"
+requires-python = ">=3.14"
 keywords = ["sonos", "lastfm", "scrobbler", "music"]
 classifiers = [
     "Development Status :: 4 - Beta",
     "Intended Audience :: End Users/Desktop",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.14",
     "Topic :: Multimedia :: Sound/Audio",
 ]
 dependencies = [
-    "soco>=0.27.0",
-    "pylast>=5.1.0",
-    "python-dotenv>=1.0.0",
-    "typer>=0.9.0",
-    "rich>=13.7.0",
-    "keyring>=24.3.0",
+    "soco>=0.30.12",
+    "pylast>=6.0.0",
+    "python-dotenv>=1.2.1",
+    "typer>=0.20.0",
+    "rich>=14.2.0",
+    "keyring>=25.6.0",
 ]
 
 [project.urls]
@@ -49,7 +48,7 @@ include = [
 packages = ["src/sonos_lastfm"]
 
 [tool.mypy]
-python_version = "0.1.5"
+python_version = "3.14"
 strict = true
 warn_return_any = true
 warn_unused_configs = true
@@ -65,10 +64,11 @@ warn_unreachable = true
 show_error_codes = true
 
 [tool.ruff]
-target-version = "py312"
+target-version = "py314"
 line-length = 88
 lint.select = ["ALL"]
 lint.ignore = ["PLW0603"]
+preview = true
 
 [tool.ruff.lint.per-file-ignores]
 "__init__.py" = ["F401"]
