@@ -1,6 +1,11 @@
+# Copyright (c) 2025 Denis Moskalets
+# Licensed under the MIT License.
+
+"""Configuration helpers for Sonos Last.fm scrobbler."""
+
 import os
 from pathlib import Path
-from typing import List, Optional
+from typing import Final
 
 from dotenv import load_dotenv
 
@@ -8,7 +13,12 @@ from dotenv import load_dotenv
 load_dotenv()
 
 
-def validate_config() -> Optional[List[str]]:
+MIN_THRESHOLD_PERCENT: Final[float] = 0.0
+MAX_THRESHOLD_PERCENT: Final[float] = 100.0
+DEFAULT_THRESHOLD_PERCENT: Final[float] = 25.0
+
+
+def validate_config() -> list[str] | None:
     """Validate required environment variables.
 
     Returns:
@@ -22,20 +32,25 @@ def validate_config() -> Optional[List[str]]:
     ]
 
     missing_vars = [var for var in required_vars if not os.getenv(var)]
-    return missing_vars if missing_vars else None
+    return missing_vars or None
 
 
-def get_config():
+def get_config() -> dict[str, object]:
     """Get configuration values, validating them first.
 
     Raises:
         ValueError: If required environment variables are missing
+
+    Returns:
+        A dictionary containing validated configuration settings.
     """
     if missing := validate_config():
-        raise ValueError(
-            f"Missing required environment variables: {', '.join(missing)}\n"
-            "Please set them in your .env file",
+        missing_vars = ", ".join(missing)
+        error_message = (
+            "Missing required environment variables: "
+            f"{missing_vars}\nPlease set them in your .env file"
         )
+        raise ValueError(error_message)
 
     return {
         # Last.fm API credentials
@@ -50,7 +65,11 @@ def get_config():
         ),  # seconds
         # Get and validate scrobble threshold percentage
         "SCROBBLE_THRESHOLD_PERCENT": min(
-            max(float(os.getenv("SCROBBLE_THRESHOLD_PERCENT") or "25"), 0), 100
+            max(
+                float(os.getenv("SCROBBLE_THRESHOLD_PERCENT") or "25"),
+                MIN_THRESHOLD_PERCENT,
+            ),
+            MAX_THRESHOLD_PERCENT,
         ),
         # Data storage paths
         "DATA_DIR": Path("./data"),
@@ -70,9 +89,11 @@ SPEAKER_REDISCOVERY_INTERVAL = int(
 )  # seconds
 
 # Get and validate scrobble threshold percentage
-SCROBBLE_THRESHOLD_PERCENT = float(os.getenv("SCROBBLE_THRESHOLD_PERCENT") or "25")
-if not 0 <= SCROBBLE_THRESHOLD_PERCENT <= 100:
-    SCROBBLE_THRESHOLD_PERCENT = 25
+SCROBBLE_THRESHOLD_PERCENT = float(
+    os.getenv("SCROBBLE_THRESHOLD_PERCENT") or str(DEFAULT_THRESHOLD_PERCENT),
+)
+if not MIN_THRESHOLD_PERCENT <= SCROBBLE_THRESHOLD_PERCENT <= MAX_THRESHOLD_PERCENT:
+    SCROBBLE_THRESHOLD_PERCENT = DEFAULT_THRESHOLD_PERCENT
 
 # Data storage paths
 DATA_DIR = Path("./data")
