@@ -1,7 +1,10 @@
+# Copyright (c) 2025 Denis Moskalets
+# Licensed under the MIT License.
+
 """Sonos Last.fm scrobbler package."""
 
 from .cli import main
 from .sonos_lastfm import SonosScrobbler
 
-__version__ = "0.1.5"
-__all__ = ["main", "SonosScrobbler"]
+__version__ = "0.1.7"
+__all__ = ["SonosScrobbler", "main"]
